Move QueryParamsProxy to its own class
