+++ lib/streamlit/runtime/state/session_state_proxy.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Any, Dict, Iterator, List, MutableMapping
+from typing import Any, Dict, Iterator, MutableMapping
 
 from typing_extensions import Final
 
@@ -20,10 +20,6 @@ from streamlit import logger as _logger
 from streamlit import runtime
 from streamlit.runtime.metrics_util import gather_metrics
 from streamlit.runtime.state.common import require_valid_user_key
-from streamlit.runtime.state.query_params import (
-    QueryParams,
-    _missing_key_error_message_query_params,
-)
 from streamlit.runtime.state.safe_session_state import SafeSessionState
 from streamlit.runtime.state.session_state import SessionState
 from streamlit.type_util import Key
@@ -57,76 +53,6 @@ def get_session_state() -> SafeSessionState:
                 )
         return SafeSessionState(SessionState(), lambda: None)
     return ctx.session_state
-
-
-def get_query_params() -> QueryParams:
-    from streamlit.runtime.scriptrunner import get_script_run_ctx
-
-    ctx = get_script_run_ctx()
-
-    if ctx is None:
-        return QueryParams()
-    return get_session_state()._state._query_params
-
-
-class QueryParamsProxy(MutableMapping[Key, Any]):
-    """A stateless singleton that proxies `st.query_params` interactions
-    to the current script thread's QueryParams instance. It stores str keys with str and List[str] values.
-    """
-
-    def __iter__(self) -> Iterator[Any]:
-        return iter(get_query_params())
-
-    def __len__(self) -> int:
-        return len(get_query_params())
-
-    def __getitem__(self, key: str) -> str:
-        return get_query_params()[key]
-
-    def __getattr__(self, key: str) -> str:
-        try:
-            return get_query_params()[key]
-        except KeyError:
-            raise AttributeError(_missing_key_error_message_query_params(key))
-
-    def __delitem__(self, key: str) -> None:
-        del get_query_params()[key]
-
-    def __delitem__(self, key: str) -> None:
-        try:
-            del get_query_params()[key]
-        except KeyError:
-            raise AttributeError(_missing_key_error_message_query_params(key))
-
-    def __setattr__(self, key: str, value: Any) -> None:
-        try:
-            get_query_params()[key] = value
-        except KeyError:
-            raise AttributeError(_missing_key_error_message_query_params(key))
-
-    def __setitem__(self, key: str, value: str) -> None:
-        get_query_params()[key] = value
-
-    def get_all(self, key: str) -> List[str]:
-        return get_query_params().get_all(key)
-
-    def __contains__(self, key: str) -> bool:
-        return key in get_query_params()
-
-    def __len__(self) -> int:
-        return len(get_query_params())
-
-    def clear(self) -> None:
-        get_query_params().clear()
-
-    def __delattr__(self, key: str) -> None:
-        try:
-            del get_query_params()[key]
-        except KeyError:
-            raise AttributeError(_missing_key_error_message_query_params(key))
-
-    def get(self, key: str, default: str = None):
-        return get_query_params().get(key, default)
 
 
 class SessionStateProxy(MutableMapping[Key, Any]):
