+++ lib/tests/streamlit/runtime/state/session_state_proxy_test.py
@@ -21,12 +21,7 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from streamlit.errors import StreamlitAPIException
-from streamlit.runtime.state import (
-    QueryParamsProxy,
-    SafeSessionState,
-    SessionState,
-    SessionStateProxy,
-)
+from streamlit.runtime.state import SafeSessionState, SessionState, SessionStateProxy
 from streamlit.runtime.state.common import (
     GENERATED_WIDGET_ID_PREFIX,
     require_valid_user_key,
@@ -139,58 +134,3 @@ class SessionStateProxyAttributeTests(unittest.TestCase):
     def test_setattr(self):
         self.session_state_proxy.corge = "grault2"
         assert self.session_state_proxy.corge == "grault2"
-
-
-class TestQueryParamsProxy:
-    @pytest.fixture
-    def proxy(self):
-        """Fixture to create a QueryParamsProxy instance."""
-        proxy = QueryParamsProxy()
-        proxy["test"] = "value"
-        return proxy
-
-    def test_get_item(self, proxy):
-        assert proxy["test"] == "value"
-
-    def test_set_item(self, proxy):
-        assert "test" in proxy
-
-    def test_del_item(self, proxy):
-        del proxy["test"]
-        assert "test" not in proxy
-
-    def test_len(self, proxy):
-        assert len(proxy) == 1
-
-    def test_iter(self, proxy):
-        keys = list(iter(proxy))
-        assert keys == ["test"]
-
-    def test_contains(self, proxy):
-        assert "test" in proxy
-
-    def test_clear(self, proxy):
-        proxy.clear()
-        assert len(proxy) == 0
-
-    def test_get(self, proxy):
-        assert proxy.get("test") == "value"
-
-    def test_get_default(self, proxy):
-        assert proxy.get("nonexistent", "bob") == "bob"
-
-    def test_get_all(self, proxy):
-        proxy["test"] = ["value1", "value2"]
-        assert proxy.get_all("test") == ["value1", "value2"]
-
-    def test_getattr(self, proxy):
-        proxy["test"] = "value"
-        assert proxy.test == "value"
-
-    def test_setattr(self, proxy):
-        proxy.test = "value"
-        assert proxy["test"] == "value"
-
-    def test_delattr(self, proxy):
-        del proxy.test
-        assert "test" not in proxy
